fix apk path bug

+++ extractFunc.sh
@@ -15,8 +15,6 @@ fi
 BASE_DIR="./$PACKAGE_NAME"
 LOCAL_APK="$BASE_DIR/${PACKAGE_NAME}.apk"
 
-mkdir -p "$BASE_DIR"
-
 if [ -f "$LOCAL_APK" ]; then
     echo "📂 APK가 이미 존재함."
 else
@@ -29,26 +27,34 @@ else
     echo "📦 패키지 경로 찾는 중..."
     PKG=$(adb shell "pm list packages" | grep "$PACKAGE_NAME" | head -n 1 | cut -d: -f2)
     PACKAGE_NAME=$PKG
-    echo "📦 패키지 이름: $PACKAGE_NAME"
+    BASE_DIR="./$PACKAGE_NAME"
+    LOCAL_APK="$BASE_DIR/${PACKAGE_NAME}.apk"
 
-    APK_PATH=$(adb shell pm path "$PACKAGE_NAME" | grep '^package:' | grep 'base.apk' | sed 's/package://')
+    if [ -f "$LOCAL_APK" ]; then
+        echo "📂 APK가 이미 존재함."
+    else
+        mkdir -p "$BASE_DIR"
+        echo "📦 패키지 이름: $PACKAGE_NAME"
 
-    if [ -z "$APK_PATH" ]; then
-        echo "❌ base.apk 경로를 찾을 수 없습니다."
-        exit 1
+        APK_PATH=$(adb shell pm path "$PACKAGE_NAME" | grep '^package:' | grep 'base.apk' | sed 's/package://')
+
+        if [ -z "$APK_PATH" ]; then
+            echo "❌ base.apk 경로를 찾을 수 없습니다."
+            exit 1
+        fi
+
+        echo "✅ APK 경로: $APK_PATH"
+
+        echo "📲 APK 파일 추출 중..."
+        adb shell "cp $APK_PATH /sdcard/${PACKAGE_NAME}.apk"
+        adb pull "/sdcard/${PACKAGE_NAME}.apk" "$LOCAL_APK"
+
+        if [ ! -f "$LOCAL_APK" ]; then
+            echo "❌ APK 파일 추출 실패"
+            exit 1
+        fi
+        echo "✅ APK 파일 추출 완료: $LOCAL_APK"
     fi
-
-    echo "✅ APK 경로: $APK_PATH"
-
-    echo "📲 APK 파일 추출 중..."
-    adb shell "cp $APK_PATH /sdcard/${PACKAGE_NAME}.apk"
-    adb pull "/sdcard/${PACKAGE_NAME}.apk" "$LOCAL_APK"
-
-    if [ ! -f "$LOCAL_APK" ]; then
-        echo "❌ APK 파일 추출 실패"
-        exit 1
-    fi
-    echo "✅ APK 파일 추출 완료: $LOCAL_APK"
 fi
 
 echo
